backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
python-dotenv==1.0.1
anthropic==0.28.0
sympy==1.12
sqlalchemy==2.0.30
psycopg2-binary==2.9.6
pydantic==2.7.1
pydantic-settings==2.3.0
python-multipart==0.0.9
httpx==0.27.0